requirements.txt
langchain
langchain-community
langchain-ollama
langchain-text-splitters
chromadb
duckdb
python-dotenv
tqdm
